% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasstimp.r
\name{perform_imputation}
\alias{perform_imputation}
\title{Basic imputation function}
\usage{
perform_imputation(gwas, pc, thresh = 0.9, eval_frac = 0.5)
}
\arguments{
\item{gwas}{A data frame with columns BETA = vector of effect estimates in the same order as ld_matrix and with NAs for variants that need to be imputed;
SE = as with BETA but for available standard errors, EAF = allele frequencies (no missing values allowed, so use reference panel if there are missing values)}

\item{pc}{list of values and vectors from PCA of LD matrix}

\item{thresh}{Fraction of variance of LD matrix to use for projection.}

\item{eval_frac}{Fraction of largest betas to use to test for agreement of imputation and observed betas}
}
\value{
A list with the following elements:
\itemize{
\item gwas: The input data frame with the imputed values added
\item z_adj_coef1: The adjustment factor for the effect sizes
\item se_adj_coef1: The adjustment factor for the standard errors
\item b_cor: The correlation between the true and imputed effect sizes - this is critical for evaluation of the performance of the imputation,
it should be close to 1 e.g > 0.7 would be a reasonable threshold
\item se_cor: The correlation between the true and imputed standard errors
}
}
\description{
Basic imputation function
}
