% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.r
\name{simulate_ss}
\alias{simulate_ss}
\title{Simulate summary statistics from genotype data}
\usage{
simulate_ss(X, af, ncause, sigmag, seed = 1234)
}
\arguments{
\item{X}{Genotype matrix (individuals x SNPs)}

\item{af}{Allele frequencies}

\item{ncause}{Number of causal variants}

\item{sigmag}{Standard deviation of effect sizes}

\item{seed}{Random seed for reproducibility}
}
\value{
A tibble with simulated summary statistics: betahat, b (true effect sizes), se, zhat, pval, af
}
\description{
Simulate summary statistics from genotype data
}
