% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasstimp.r
\name{filter_imputation_results}
\alias{filter_imputation_results}
\title{filter_imputation_results: remove results with over-inflated p-vlaues}
\usage{
filter_imputation_results(gwas, ld_matrix, min_bp, max_bp)
}
\arguments{
\item{imputed}{gwas}

\item{ld}{correlation matrix

First, filters the imputation results to inside the range of the original gwas
Second, finds imputed variants with low pvalues, non-imputed rows that are correlated
If imputed variant is more significant the non-imputed rows, drop that variant.
Also, if the imputed variant has an illegal value, drop that variant.}
}
\value{
filtered gwas
}
\description{
filter_imputation_results: remove results with over-inflated p-vlaues
}
